refactor: article route refactor, renamed, news creator tool routes file added. slack routes fix
